feat(domain) : add features to Lotto
로또 기능 추가
 - 번호 정렬 기능
 - 당첨된 갯수 반환 기능
 - 등수 반환
 - 출력 메소드 오버라이딩
